fix broken item rendering and refactor fns out of main mod file into utils and handlers
