+++ src/main/java/com/lothrazar/samsapples/UtilItemRenderer.java
@@ -1,0 +1,27 @@
+package com.lothrazar.samsapples;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.client.renderer.texture.TextureMap;
+import net.minecraft.client.resources.model.IBakedModel;
+import net.minecraft.item.ItemStack;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
+
+public class UtilItemRenderer
+{
+	@SideOnly(Side.CLIENT)
+	public static void renderItemAt(ItemStack stack, int x, int y, int dim)
+	{
+		//first get texture from item stack
+		IBakedModel iBakedModel = Minecraft.getMinecraft().getRenderItem().getItemModelMesher().getItemModel(stack);
+		TextureAtlasSprite textureAtlasSprite = Minecraft.getMinecraft().getTextureMapBlocks().getAtlasSprite(iBakedModel.getTexture().getIconName());
+	 
+		
+		Minecraft.getMinecraft().getTextureManager().bindTexture(TextureMap.locationBlocksTexture);
+       
+        if(Minecraft.getMinecraft().currentScreen != null)
+        	Minecraft.getMinecraft().currentScreen.drawTexturedModalRect(x,y,textureAtlasSprite, dim,dim);
+        
+	}
+}
